move repo info fields to their own files
