+++ src/info/repo/mod.rs
@@ -1,0 +1,13 @@
+pub mod author;
+pub mod commits;
+pub mod contributors;
+pub mod created;
+pub mod head;
+pub mod last_change;
+pub mod license;
+pub mod loc;
+pub mod pending;
+pub mod project;
+pub mod size;
+pub mod url;
+pub mod version;
